--- App.js
@@ -1,14 +1,0 @@
-import List from './List';
-
-function App(props) {
-  return (
-    <div>
-
-      <List></List>
-
-    </div>
-  );
-}
-export default App;
-
-
