+++ src/App.js
@@ -1,0 +1,15 @@
+import { useState } from "react";
+import List from './List';
+import "./App.css";
+function App(_props) {
+  return (
+    <div className='app'>
+
+      <List></List>
+
+    </div>
+  );
+}
+export default App;
+
+
