+++ src/List.js
@@ -1,13 +1,14 @@
-import ListItem from "./ListItem"
 import { useState } from "react";
+import ListItem from "./ListItem"
 
 function List() {
     let text
-
     const [list, setList] = useState([]);
+    
 const f =list.map((element,index)=>{
-    // return <li>{element}</li>
-return <ListItem elements={element} index={index}/>
+     return <li>{element}
+      <ListItem elements={element} index={index}/></li>
+
 })
 
     const addItem = (e) => {
@@ -27,17 +28,18 @@ return <ListItem elements={element} index={index}/>
     return (
         <div>
             <h1>to do </h1>
+            {f}
+           
             <h2>what needs to be done</h2>
             <input type="text" onChange={changeText} />
             <button onClick={addItem}>Add</button>
-            {f}
+           
         </div>
 
 
     )
 }
-export default List
-
+export default List;
